RF: added test for popeye.prf and popeye.stimulus and removed redundancy from previously used tests
